+++ src/StaticLS/SrcFiles.hs
@@ -1,0 +1,10 @@
+module StaticLS.SrcFiles (
+    SrcFilePath,
+    srcDirs,
+) where
+
+type SrcFilePath = FilePath
+
+-- TODO: make this configurable (use cabal?)
+srcDirs :: [FilePath]
+srcDirs = ["src/", "lib/", "app/", "test/"]