+++ test/SpecHook.hs
@@ -11,8 +11,7 @@ import Prelude
 hook :: Spec -> Spec
 hook = do
     beforeAll
-        ( indexHieFiles
-            >> removeIfExists testHieDbDir
+        ( removeIfExists testHieDbDir >> indexHieFiles >> removeIfExists testHieDbDir
         )
   where
     removeIfExists :: FilePath -> IO ()